Refactor: Extract workspace utility functions to separate module (Phase 1)
Extract pure utility functions from main.ts into a focused workspace-utils module.
This is Phase 1 of the incremental main.ts refactoring strategy (issue #237).

## Changes

**New Module: src/lib/workspace-utils.ts (77 lines)**
- expandTildePath() - Path expansion with tilde support
- showWorkspaceError() - UI error display helper
- clearWorkspaceError() - UI error clearing helper

**Updated: src/main.ts**
- Import utility functions from workspace-utils
- Remove 44 lines of duplicated code
- Reduced from 2,196 to 2,152 lines

## Benefits

- **Zero Risk**: Pure functions with no state dependencies
- **Immediate Value**: 44-line reduction in main.ts
- **Better Organization**: Workspace utilities now in dedicated module
- **Improved Documentation**: Functions now have JSDoc comments
- **Easier Testing**: Utilities can be tested in isolation

## Testing

- ✅ TypeScript compilation passes
- ✅ Biome linting passes
- ✅ Biome formatting passes
- ✅ Frontend build passes
- ✅ No behavioral changes

Part of issue #237 - Phase 1 of incremental refactoring strategy

Related: #237
